+++ src/main/java/com/example/event_manager/service/BillingRaportServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.event_manager.utils;
+package com.example.event_manager.service;
 
 import com.example.event_manager.form.BillingForm;
 import com.example.event_manager.model.BillingRaportSchema;
@@ -19,36 +19,32 @@ import org.apache.fop.apps.FOUserAgent;
 import org.apache.fop.apps.Fop;
 import org.apache.fop.apps.FopFactory;
 import org.apache.fop.apps.MimeConstants;
+import org.springframework.stereotype.Service;
 import org.springframework.util.ResourceUtils;
 
-public class BillingRaport {
+@Service
+public class BillingRaportServiceImpl implements BillingRaportService {
 
   private static final String styleXslFileLocation = "classpath:static/newStyle.xsl";
   private static File styleFile;
-  private final BillingRaportSchema billingRaportSchema;
-  private static String billingsXml;
 
 
-  public BillingRaport(final BillingRaportSchema billingRaportSchema)
+  public String generateXmlForBillingRaportSchema(BillingRaportSchema brs)
       throws FileNotFoundException {
-    this.billingRaportSchema = billingRaportSchema;
     styleFile = ResourceUtils.getFile(styleXslFileLocation);
-  }
-
-  private void generateXmlForBillingRaportSchema() {
     final XStream xStream = new XStream();
     xStream.alias("BillingRaportSchema", BillingRaportSchema.class);
     xStream.alias("billing", BillingForm.class);
     xStream.addImplicitCollection(BillingRaportSchema.class, "billings");
-    billingsXml = xStream.toXML(billingRaportSchema);
+    return xStream.toXML(brs);
 
   }
 
-  public byte[] convertBillingRaportToByteStream()
+  public byte[] convertBillingRaportSchemaToByteStream(BillingRaportSchema brs)
       throws IOException, FOPException, TransformerException {
-    generateXmlForBillingRaportSchema();
+    String xml = generateXmlForBillingRaportSchema(brs);
     final File xsltFile = styleFile;
-    final StreamSource xmlSource = new StreamSource(new StringReader(billingsXml));
+    final StreamSource xmlSource = new StreamSource(new StringReader(xml));
     final FopFactory fopFactory = FopFactory.newInstance(new File(".").toURI());
     final FOUserAgent foUserAgent = fopFactory.newFOUserAgent();
     final ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
