+++ src/main/java/com/example/event_manager/service/BillingServiceImpl.java
@@ -45,4 +45,5 @@ public class BillingServiceImpl implements BillingService {
     return billingMapper.toDto(findById(id));
   }
 
+
 }
